backend/skillnest/requirements.txt
absl-py==2.2.2
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.3.0
billiard==4.2.1
cachetools==5.5.2
celery==5.5.3
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudpickle==3.1.1
colorama==0.4.6
contourpy==1.3.1
coredis==4.20.0
cron-descriptor==1.4.5
crow-client==0.3.6.dev12
cryptography==45.0.5
cycler==0.12.1
defusedxml==0.7.1
Deprecated==1.2.18
distro==1.9.0
dj-rest-auth==7.0.1
Django==5.2.1
django-allauth==65.10.0
django-celery-beat==2.8.1
django-cors-headers==4.7.0
django-timezone-field==7.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
dm-tree==0.1.9
dnspython==2.7.0
docstring_parser==0.16
email_validator==2.2.0
fhaviary==0.18.3
fhlmi==0.26.0
filelock==3.18.0
fonttools==4.57.0
frozenlist==1.5.0
fsspec==2025.3.2
google-auth==2.40.3
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.30.1
idna==3.10
importlib_metadata==8.6.1
Jinja2==3.1.6
jiter==0.9.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
kombu==5.5.4
ldp==0.26.0
limits==4.6
litellm==1.65.4.post1
MarkupSafe==3.0.2
matplotlib==3.10.1
multidict==6.3.2
networkx==3.4.2
numpy==2.2.4
oauthlib==3.3.1
openai==1.70.0
packaging==24.2
pandas==2.2.3
pillow==11.1.0
prompt_toolkit==3.0.51
propcache==0.3.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.2
pydantic_core==2.33.1
PyJWT==2.9.0
Pympler==1.1
pyparsing==3.2.3
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python3-openid==3.2.0
pytz==2025.2
pywin32==310
PyYAML==6.0.2
redis==6.2.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==1.3.0
rpds-py==0.24.0
rsa==4.9.1
scipy==1.15.2
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
tenacity==9.1.2
tiktoken==0.9.0
tokenizers==0.21.1
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.1
tzdata==2025.2
urllib3==2.3.0
usearch==2.17.5
vine==5.1.0
wcwidth==0.2.13
wheel==0.45.1
wrapt==1.17.2
yarl==1.19.0
zipp==3.21.0
